🔨 Modularizar código del servidor
